Split up server routing
